Updates to p4app-exercises/p4runtime.p4app files
With these changes, the README is nearly identical to
exercises/p4runtime/README.md file, with only the updates necessary to
make it accurate for the p4app version of the code.

There may still be a few fixes needed to get it completely correct,
because when I ran `make solutioncontroller`, I saw only 3 out of the
4 counter values incrementing over time, but suspect that the desired
behavior is for all 4 of those counters to be incrementing?