Several proposed changes to v1model standard_metadata for cheat sheet
Remove several fields that are deprecated in the v1model.p4 include file,
or will be soon:

+ drop
+ recirculate_port
+ clone_spec
+ lf_field_list
+ resubmit_flag
+ recirculate_flag

Add mention of more detailed documentation in behavioral-model
repository, for those that want to know more about how each field
behaves.

Group the fields by those that are only intended to be read in a P4
program, versues read/write, and which are only intended to be useful
in ingress, egress, or both.