2022 may updates to vm ubuntu 20.04 (#470)
* Put latest commits for p4lang repos as of 2022-May-01 in build scripts

* Updates to steps to build behavioral-model for 2022-May
The parts of these changes on how to build simple_switch_grpc are due
to changes made in the behavioral-model recently, perhaps in 2022-Apr.

The removing of the applying of the patch is from older changes in
behavioral-model around 2022-Jan.  The patch is no longer needed.

* Correction to script changes in previous commit

* Add --with-thrift option to configure command for behavioral-model
I believe that without that, the simple_switch_grpc process is built
without support for listening on the Thrift port (default value of
port is 9090).

Co-authored-by: Andy Fingerhut <john.andy.fingerhut@intel.com>
Signed-off-by: Andy Fingerhut <andy_fingerhut@alum.wustl.edu>