Remove part of README install instructions that are obsolete (#494)
They refer to bash script names that do not exist in the
vm-ubuntu-20.04 directory, and it is not straightforward and quick to
update those instructions to use scripts in the vm-ubuntu-20.04
directory without making a fair number of changes, and probably
several hours of testing those changes, that I am not prepared to make
any time soon.

The instructions _might_ be correct for the vm directory when starting
with an Ubuntu 16.04 Linux system, but that version of Ubuntu is no
longer supported, and I do not think we should be recommending its
use.

Signed-off-by: Andy Fingerhut <andy_fingerhut@alum.wustl.edu>