Basic tunnel exercise fixes (#625)
* Support negative-valued action parameters
by converting them in the Python helper code into a positive value
whose bit pattern is the 2's complement representation of the negative
value.

* Change send.py for exercise basic_tunnel so its README is correct
The README said to expect a packet with a TCP header at the receiver,
but send.py was not adding a TCP header to the packet when the
`--dst_id` command line option was specified.

---------

Co-authored-by: jafinger <jafinger@cisco.com>
Signed-off-by: Andy Fingerhut <andy.fingerhut@gmail.com>